ext4_xattr: convert __s32 to size_t for ext4_xattr_ref::inode/block_size_rem

+++ Ext3Fsd/include/linux/ext4_xattr.h
@@ -147,8 +147,8 @@ struct ext4_xattr_ref {
 
 	BOOL   dirty;
 	size_t ea_size;
-	__s32 inode_size_rem;
-	__s32 block_size_rem;
+	size_t inode_size_rem;
+	size_t block_size_rem;
 	PEXT2_VCB fs;
 
 	void *iter_arg;
